+++ core/src/internal/index.ts
@@ -26,6 +26,7 @@ import * as connectionHolder from './connection-holder'
 import * as txConfig from './tx-config'
 import * as transactionExecutor from './transaction-executor'
 import * as connectivityVerifier from './connectivity-verifier'
+import * as logger from './logger'
 
 export {
   util,
@@ -36,5 +37,6 @@ export {
   connectionHolder,
   txConfig,
   transactionExecutor,
-  connectivityVerifier
+  connectivityVerifier,
+  logger
 }
